Rename/Move GeoTrackingButton to own component
